requirements.txt

# requirements.txt (actualizado)
# Recomendado: Python 3.10–3.12 (TensorFlow no soporta aún 3.13 de forma estable)
tensorflow==2.20.0
streamlit>=1.37.0
pillow
numpy
